Use native crypto support on Windows
As we only need HMAC and RNG there isn't much OS-specific code needed,
and for applications that don't already depend on OpenSSL or GnuTLS it
simplifies the build process quite a lot.

Note that we use the vintage WinCrypt APIs as the Cryptography Next
Generation APIs are only available on Vista and newer.