agent: Freeze list of local addresses when gathering local candidates
With this patch, we establish the list of local addresses when gathering
local candidates, and only there. We consider this list will remain
constant during the lifetime of the agent, the same way we keep the same
list of local host candidates.

Reevaluating this list of local addresses while the agent is running may
have side effects when computing stun priority of pairs, whose a part of
this priority is based on the local address rank in this list. This rank
may change, if network interfaces are added or removed while the agent
is running.

We also replace the list of local IP addresses by the list of local
NiceAddress addresses, which is used when gathering local candidate,
when local candidate are manually added, and whose lifetime can
extended.  This list provides us with the same information, and can be
used to compute a fixed stun priority per address as well.