agent: freeze the list of local ip addresses when gathering local candidates
With this patch, we establish the list of local ip addresses when
gathering local candidates, and only there. We consider this list will
remain constant during the lifetime of the agent, the same way we keep
the same list of local host candidates.

Reevaluating this list of ip addresses while the agent is running may
have side effects when computing stun priority of pairs, whose a part of
this priority is based on the ip address rank in this list. This rank
may change, if network interfaces are added or removed while the agent is
running.